+++ .storybook-openbrush/main.js
@@ -1,0 +1,25 @@
+/** @type {import('@storybook/html-webpack5').StorybookConfig} */
+const config = {
+  stories: [
+    '../src/openbrush/atoms/**/*.stories.js',
+    '../src/openbrush/molecules/**/*.stories.js',
+    '../src/openbrush/organisms/**/*.stories.js',
+  ],
+  addons: ['@storybook/addon-essentials'],
+  framework: {
+    name: '@storybook/html-webpack5',
+    options: {},
+  },
+  webpackFinal: async (config) => {
+    config.devServer = {
+      ...config.devServer,
+      client: {
+        ...config.devServer?.client,
+        webSocketURL: 'auto://0.0.0.0:0/ws',
+      },
+    };
+    return config;
+  },
+};
+
+export default config;